flattened the package hierarchy
this'll help in future to avoid packages names collisions
