+++ internal/sip/codes.go
@@ -1,4 +1,4 @@
-package status
+package sip
 
 type (
 	Code   uint16
