+++ internal/sip/states.go
@@ -1,4 +1,4 @@
-package parser
+package sip
 
 type parserState int
